Improve error message on failed root filesystem access and add hint to README
The AFC2 service required to access the root filesystem is non-standard.
It is installed by older jailbreaking software but for instance not by
blackra1n for devices like the 3GS. Now users get a better hint at why
their attempts to use ifuse fail with the --root switch. This also moves
the print to stderr where it actually can be seen by users.

[#107 state:resolved]